ffmpeg_to_fox: Simplify user input and exception handling using pathlib
Handling of the worse user input faults to catch unwanted side effects
is not enough and is not simple to understand.

Used Pathlib to rewrite the input checking section to be:
Input path - existing, directory, read permissions
Output path - existing and empty or non-existing, write permissions

Exceptions are always thrown up because we do not know what exceptions
rosbag2 library can produce. Most of the exceptions we can expect are
the same RuntimeError with different exception message. Rosbag2 is mostly
a wrapper around storage plugin. Storage plugins are usually c/c++ library
and for example most of mcap library errors are propagated up
as python RuntimeError exception.